Backend/requirements.txt
annotated-types
Flask==2.0.3
gunicorn==20.1.0
asttokens
cachetools
certifi
charset-normalizer
decorator
executing
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
grpcio
grpcio-status
httplib2
idna
ipython
jedi
matplotlib-inline
numpy
opencv-python
parso
pexpect
pillow
prompt_toolkit
proto-plus
protobuf
ptyprocess
pure-eval
pyasn1
pyasn1_modules
pydantic
pydantic_core
Pygments
pyparsing
requests
rsa
six
stack-data
tqdm
traitlets
typing_extensions
uritemplate
urllib3
wcwidth
Pillow
opencv-python


# pip install --upgrade google-generativeai google-cloud-aiplatform
